refactor: change logic v1 for SwapButton (ActionsButtons)
